+++ R/get_common_dataset_type.R
@@ -9,38 +9,35 @@
 get_common_dataset_type <- function(data_source, vec_dataset_type, verbose = TRUE) {
   .data <- rlang::.data
 
-  assertthat::assert_that(
+  assertthat_cli(
     inherits(data_source, "tbl"),
-    msg = "data_source must be a class of `tbl`"
+    msg = "{.arg data_source} must be a class of {.cls tbl}.",
+    verbose = verbose
   )
 
-  assertthat::assert_that(
+  assertthat_cli(
     "dataset_type" %in% colnames(data_source),
-    msg = paste(
-      "The dataset does not contain `dataset_type` columns",
-      "during the selection of common dataset type.",
-      "Please check the column names of the dataset."
-    )
+    msg = "The dataset does not contain {.code dataset_type} column during the selection of common dataset type. Please check the column names of the dataset.",
+    verbose = verbose
   )
 
-  assertthat::assert_that(
+  assertthat_cli(
     is.character(vec_dataset_type),
-    msg = "`vec_dataset_type` must be a character vector"
+    msg = "{.arg vec_dataset_type} must be a character vector.",
+    verbose = verbose
   )
 
-  assertthat::assert_that(
+  assertthat_cli(
     all(vec_dataset_type %in% c(
       "vegetation_plot", "fossil_pollen_archive", "traits", "gridpoints"
     )),
-    msg = paste(
-      "The `vec_dataset_type` must be one of the following:",
-      "`vegetation_plot`, `fossil_pollen_archive`, `traits`, `gridpoints`"
-    )
+    msg = "{.arg vec_dataset_type} must be one of {.code vegetation_plot}, {.code fossil_pollen_archive}, {.code traits}, or {.code gridpoints}.",
+    verbose = verbose
   )
 
-  assertthat::assert_that(
+  assertthat_cli(
     inherits(verbose, "logical"),
-    msg = "`verbose` must be a logical value"
+    msg = "{.arg verbose} must be a logical value."
   )
 
   vec_present_dataset_type <-
@@ -58,13 +55,11 @@ get_common_dataset_type <- function(data_source, vec_dataset_type, verbose = TRU
     if (
       isTRUE(verbose)
     ) {
-      message(
-        paste(
-          "The data does not contain the all dataset types specified in",
-          "`vec_dataset_type`.", "\n",
-          "Changing `vec_dataset_type` to the dataset types",
-          "present in the data as:",
-          paste(vec_common_dataset_type, collapse = ", ")
+      cli::cli_alert_warning(
+        stringr::str_c(
+          "The data does not contain all dataset types specified in {.arg vec_dataset_type}. ",
+          "Changing {.arg vec_dataset_type} to the dataset types present in the data as: ",
+          stringr::str_c(vec_common_dataset_type, collapse = ", ")
         )
       )
     }
@@ -81,10 +76,9 @@ get_common_dataset_type <- function(data_source, vec_dataset_type, verbose = TRU
     if (
       isTRUE(verbose)
     ) {
-      message(
-        paste(
-          "The selected dataset type will not be filtered:",
-          paste(dataset_type_missing, collapse = ", ")
+      cli::cli_alert_warning(
+        stringr::str_glue(
+          "The selected dataset type will not be filtered: {stringr::str_c(dataset_type_missing, collapse = ', ')}"
         )
       )
     }