+++ R/select_abiotic_var_by_name.R
@@ -7,72 +7,53 @@
 select_abiotic_var_by_name <- function(con = NULL, sel_var_name = NULL) {
   .data <- rlang::.data
 
-  assertthat::assert_that(
+  assertthat_cli(
     inherits(con, "vault_pipe"),
-    msg = paste(
-      "`con` must be a class of `vault_pipe`",
-      "Use `open_vault()` to create a connection"
-    )
+    msg = "{.arg con} must be a {.cls vault_pipe} object. Use {.fn open_vault} to create a connection"
   )
 
-  assertthat::assert_that(
+  assertthat_cli(
     all(names(con) %in% c("data", "db_con")),
-    msg = paste(
-      "con must have `data` and `db_con`",
-      "Use `open_vault()` to create a connection"
-    )
+    msg = "{.arg con} must have {.code data} and {.code db_con} elements. Use {.fn open_vault} to create a connection"
   )
 
   sel_data <- con$data
 
-  assertthat::assert_that(
+  assertthat_cli(
     inherits(sel_data, "tbl"),
-    msg = "data must be a class of `tbl`"
+    msg = "{.code con$data} must be a {.cls tbl}"
   )
 
-  assertthat::assert_that(
+  assertthat_cli(
     "abiotic_variable_id" %in% colnames(sel_data),
-    msg = paste(
-      "The data does not contain `abiotic_variable_id` columns. Please add",
-      "`get_abiotic_data()` to the pipe before this function."
-    )
+    msg = "The data does not contain the {.code abiotic_variable_id} column. Use {.fn get_abiotic_data} before this function"
   )
 
-  assertthat::assert_that(
+  assertthat_cli(
     "dataset_type" %in% colnames(sel_data),
-    msg = paste(
-      "The data should be filtered only for `gridpoints`.",
-      "However, the does not contain `dataset_type` columns. Please add",
-      "`get_datasets()` to the pipe before this function."
-    )
+    msg = "The data does not contain the {.code dataset_type} column. Use {.fn get_datasets} before this function"
   )
 
   sel_con <- con$db_con
 
-  assertthat::assert_that(
+  assertthat_cli(
     inherits(sel_con, "SQLiteConnection"),
-    msg = "db_con must be a class of `SQLiteConnection`"
+    msg = "{.code con$db_con} must be a {.cls SQLiteConnection}"
   )
 
-  assertthat::assert_that(
+  assertthat_cli(
     "AbioticVariable" %in% DBI::dbListTables(sel_con),
-    msg = paste(
-      "AbioticVariable table does not exist in the Vault database",
-      "Make sure to connect to the correct database"
-    )
+    msg = "The {.code AbioticVariable} table does not exist in the database. Make sure to connect to the correct database"
   )
 
-  assertthat::assert_that(
+  assertthat_cli(
     "abiotic_variable_id" %in% colnames(dplyr::tbl(sel_con, "AbioticVariable")),
-    msg = paste(
-      "The AbioticVariable does not contain `abiotic_variable_id` column in the Vault database.",
-      "Make sure to connect to the correct database"
-    )
+    msg = "The {.code AbioticVariable} table does not contain the {.code abiotic_variable_id} column. Make sure to connect to the correct database"
   )
 
-  assertthat::assert_that(
+  assertthat_cli(
     is.character(sel_var_name),
-    msg = "`sel_var_name` must be a character vector"
+    msg = "{.arg sel_var_name} must be a character vector"
   )
 
   dat_res <-