+++ tests/testthat/test-get_taxa.R
@@ -75,17 +75,19 @@ testthat::test_that("number of raw taxa", {
 
 testthat::test_that("number of taxa by family", {
   test_datasets <-
-    open_vault(
-      path = paste(
-        tempdir(),
-        "example.sqlite",
-        sep = "/"
+    suppressWarnings(
+      open_vault(
+        path = paste(
+          tempdir(),
+          "example.sqlite",
+          sep = "/"
+        )
+      ) %>%
+      get_datasets() %>%
+      get_samples() %>%
+      get_taxa(
+        classify_to = "family"
       )
-    ) %>%
-    get_datasets() %>%
-    get_samples() %>%
-    get_taxa(
-      classify_to = "family"
     )
 
   test_n_taxa <-
@@ -169,4 +171,134 @@ testthat::test_that("size of a total dataset", {
   testthat::expect_equal(test_n_datasets, actual_n_datasets)
 
   DBI::dbDisconnect(con_db)
+})
+
+
+# classification_data argument ----
+testthat::test_that("classification_data tibble gives same genus result", {
+  test_con_orig <-
+    open_vault(
+      path = paste(
+        tempdir(),
+        "example.sqlite",
+        sep = "/"
+      )
+    ) %>%
+    get_datasets() %>%
+    get_samples() %>%
+    get_taxa(classify_to = "original")
+
+  data_class <-
+    get_classification_table(
+      con = test_con_orig,
+      return_raw_data = TRUE
+    )
+
+  res_with_data <-
+    suppressWarnings(
+      open_vault(
+        path = paste(
+          tempdir(),
+          "example.sqlite",
+          sep = "/"
+        )
+      ) %>%
+      get_datasets() %>%
+      get_samples() %>%
+      get_taxa(
+        classify_to = "genus",
+        classification_data = data_class
+      )
+    ) %>%
+    purrr::chuck("data") %>%
+    dplyr::distinct(taxon_id) %>%
+    dplyr::collect() %>%
+    tidyr::drop_na() %>%
+    dplyr::arrange(taxon_id) %>%
+    dplyr::pull("taxon_id")
+
+  res_without_data <-
+    suppressWarnings(
+      open_vault(
+        path = paste(
+          tempdir(),
+          "example.sqlite",
+          sep = "/"
+        )
+      ) %>%
+      get_datasets() %>%
+      get_samples() %>%
+      get_taxa(classify_to = "genus")
+    ) %>%
+    purrr::chuck("data") %>%
+    dplyr::distinct(taxon_id) %>%
+    dplyr::collect() %>%
+    tidyr::drop_na() %>%
+    dplyr::arrange(taxon_id) %>%
+    dplyr::pull("taxon_id")
+
+  testthat::expect_equal(res_with_data, res_without_data)
+})
+
+testthat::test_that("custom classification_data overrides genus mapping", {
+  test_con_orig <-
+    open_vault(
+      path = paste(
+        tempdir(),
+        "example.sqlite",
+        sep = "/"
+      )
+    ) %>%
+    get_datasets() %>%
+    get_samples() %>%
+    get_taxa(classify_to = "original")
+
+  data_class_custom <-
+    get_classification_table(
+      con = test_con_orig,
+      return_raw_data = TRUE
+    ) %>%
+    dplyr::mutate(taxon_genus = 46L)
+
+  res <-
+    suppressWarnings(
+      open_vault(
+        path = paste(
+          tempdir(),
+          "example.sqlite",
+          sep = "/"
+        )
+      ) %>%
+      get_datasets() %>%
+      get_samples() %>%
+      get_taxa(
+        classify_to = "genus",
+        classification_data = data_class_custom
+      )
+    ) %>%
+    purrr::chuck("data") %>%
+    dplyr::distinct(taxon_id) %>%
+    dplyr::collect() %>%
+    tidyr::drop_na() %>%
+    dplyr::pull("taxon_id")
+
+  testthat::expect_equal(res, 46L)
+})
+
+testthat::test_that("error when classification_data is not data.frame", {
+  testthat::expect_error(
+    open_vault(
+      path = paste(
+        tempdir(),
+        "example.sqlite",
+        sep = "/"
+      )
+    ) %>%
+      get_datasets() %>%
+      get_samples() %>%
+      get_taxa(
+        classify_to = "genus",
+        classification_data = "bad"
+      )
+  )
 })